% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{camera_orthographic}
\alias{camera_orthographic}
\title{Create an orthographic camera}
\usage{
camera_orthographic(position, target, view_height)
}
\arguments{
\item{position}{Camera position as c(x, y, z)}

\item{target}{Look-at target as c(x, y, z)}

\item{view_height}{View height in world units}
}
\value{
External pointer to Camera
}
\description{
Create an orthographic camera
}
