% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{camera_perspective}
\alias{camera_perspective}
\title{Create a perspective camera}
\usage{
camera_perspective(position, target, fov_degrees)
}
\arguments{
\item{position}{Camera position as c(x, y, z)}

\item{target}{Look-at target as c(x, y, z)}

\item{fov_degrees}{Field of view in degrees}
}
\value{
External pointer to Camera
}
\description{
Create a perspective camera
}
