% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{light_new}
\alias{light_new}
\title{Create a directional light}
\usage{
light_new(direction, intensity)
}
\arguments{
\item{direction}{Direction toward light source as c(x, y, z)}

\item{intensity}{Light intensity (default: 1.0)}
}
\value{
External pointer to Light
}
\description{
Create a directional light
}
