% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{marching_cubes}
\alias{marching_cubes}
\title{Extract isosurface using marching cubes}
\usage{
marching_cubes(volume, iso_value)
}
\arguments{
\item{volume}{Volume created with volume_new()}

\item{iso_value}{Scalar value for the isosurface}
}
\value{
External pointer to Mesh with positions, normals, and indices
}
\description{
Extract isosurface using marching cubes
}
