% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{material_solid}
\alias{material_solid}
\alias{material_solid_alpha}
\title{Create a solid color material}
\usage{
material_solid(id, rgb)

material_solid_alpha(id, rgba)
}
\arguments{
\item{id}{Material identifier string}

\item{rgb}{Color as c(r, g, b) in 0-1 range}

\item{rgba}{Color with alpha as c(r, g, b, a) in 0-1 range}
}
\value{
External pointer to Material
}
\description{
Create a solid color material
}
