% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mesh_new}
\alias{mesh_new}
\title{Create a triangle mesh}
\usage{
mesh_new(positions, indices)
}
\arguments{
\item{positions}{Flat vector of vertex positions (x0, y0, z0, x1, y1, z1, ...)}

\item{indices}{Flat vector of triangle indices (i0, i1, i2, ...)}
}
\value{
External pointer to Mesh
}
\description{
Create a triangle mesh
}
