% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{mesh_with_material}
\alias{mesh_with_material}
\alias{mesh_with_normals}
\alias{mesh_with_scalars}
\alias{mesh_triangle_count}
\alias{mesh_vertex_count}
\title{Mesh modification functions}
\usage{
mesh_with_material(mesh, material_id)

mesh_with_normals(mesh, normals)

mesh_with_scalars(mesh, scalars)

mesh_triangle_count(mesh)

mesh_vertex_count(mesh)
}
\arguments{
\item{mesh}{Mesh to modify or query}

\item{material_id}{Material identifier string}

\item{normals}{Flat vector of vertex normals}

\item{scalars}{Per-vertex scalar values}
}
\value{
New Mesh for modification functions, integer for count functions
}
\description{
Mesh modification functions
}
