% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{render_to_png}
\alias{render_to_png}
\title{Render scene to PNG}
\usage{
render_to_png(scene, width, height, background)
}
\arguments{
\item{scene}{Scene to render}

\item{width}{Image width in pixels}

\item{height}{Image height in pixels}

\item{background}{Background color as c(r, g, b, a)}
}
\value{
Raw vector containing PNG data
}
\description{
Render scene to PNG
}
