% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{save_png}
\alias{save_png}
\title{Save PNG data to file}
\usage{
save_png(png_data, filename)
}
\arguments{
\item{png_data}{Raw vector of PNG data}

\item{filename}{Output filename}
}
\value{
Invisible TRUE on success
}
\description{
Save PNG data to file
}
