% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{scene_new}
\alias{scene_new}
\title{Create a new scene}
\usage{
scene_new(camera, bounds_min, bounds_max)
}
\arguments{
\item{camera}{Camera created with camera_perspective() or camera_orthographic()}

\item{bounds_min}{Minimum bounds as c(x, y, z)}

\item{bounds_max}{Maximum bounds as c(x, y, z)}
}
\value{
External pointer to Scene
}
\description{
Create a new scene
}
